Create README.md
The code directory contains code snippets for the Manning book Deep Learning for NLP. Disclaimer: due to final stage renumbering of the code listings in the book, some listings have a slightly different number than those in the book. The chapter numbers are correct, so you should be able to find them quickly.